+++ app/src/main/java/com/aspiring_creators/aichopaicho/ui/screens/DashboardScreen.kt
@@ -48,8 +48,11 @@ import coil3.request.ImageRequest
 import coil3.request.crossfade
 import com.aspiring_creators.aichopaicho.R
 import com.aspiring_creators.aichopaicho.ui.component.ButtonComponent
+import com.aspiring_creators.aichopaicho.ui.component.DashboardContent
+import com.aspiring_creators.aichopaicho.ui.component.ErrorContent
 import com.aspiring_creators.aichopaicho.ui.component.LoadingContent
 import com.aspiring_creators.aichopaicho.ui.component.LogoTopBar
+import com.aspiring_creators.aichopaicho.ui.component.NotSignedInContent
 import com.aspiring_creators.aichopaicho.ui.component.UserProfileImage
 import com.aspiring_creators.aichopaicho.viewmodel.DashboardScreenViewModel
 import com.aspiring_creators.aichopaicho.viewmodel.data.DashboardScreenUiState
@@ -133,205 +136,4 @@ fun DashboardScreen(
             }
         }
     }
-}
-
-
-
-@Composable
-private fun NotSignedInContent(
-    onSignOut: (() -> Unit)?
-) {
-    Box(
-        modifier = Modifier.fillMaxSize(),
-        contentAlignment = Alignment.Center
-    ) {
-        Column(
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Text(
-                text = "You are not signed in",
-                style = MaterialTheme.typography.headlineSmall,
-                textAlign = TextAlign.Center
-            )
-
-            Spacer(modifier = Modifier.height(16.dp))
-
-            onSignOut?.let { signOut ->
-                ButtonComponent(
-                    R.drawable.logo_skip,
-                    "Go to Sign In",
-                    onClick = signOut
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun ErrorContent(
-    errorMessage: String,
-    onRetry: () -> Unit
-) {
-    Box(
-        modifier = Modifier.fillMaxSize(),
-        contentAlignment = Alignment.Center
-    ) {
-        Column(
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Text(
-                text = "Error: $errorMessage",
-                color = MaterialTheme.colorScheme.error,
-                textAlign = TextAlign.Center,
-                modifier = Modifier.padding(16.dp)
-            )
-
-            Spacer(modifier = Modifier.height(16.dp))
-
-            ButtonComponent(
-                R.drawable.logo_skip,
-                "Retry",
-                onClick = onRetry
-            )
-        }
-    }
-}
-
-@Composable
-private fun DashboardContent(
-    uiState: DashboardScreenUiState,
-    onSignOut: (() -> Unit)?,
-    onNavigateToProfile: (() -> Unit)?,
-    onNavigateToTransactions: (() -> Unit)?,
-    onRefresh: () -> Unit,
-    onSignOutClick: () -> Unit
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(16.dp)
-    ) {
-        // Welcome Card
-        Card(
-            modifier = Modifier.fillMaxWidth(),
-            elevation = CardDefaults.cardElevation(defaultElevation = 4.dp)
-        ) {
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(16.dp),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                // User Profile Image
-                UserProfileImage(
-                    photoUrl = uiState.user?.photoUrl.toString(),
-                    userName = uiState.user?.name,
-                    modifier = Modifier.size(64.dp)
-                )
-
-                Spacer(modifier = Modifier.width(16.dp))
-
-                // User Info
-                Column(
-                    modifier = Modifier.weight(1f)
-                ) {
-                    Text(
-                        text = "Welcome back!",
-                        style = MaterialTheme.typography.titleMedium,
-                        fontWeight = FontWeight.Normal
-                    )
-
-                    Spacer(modifier = Modifier.height(4.dp))
-
-                    Text(
-                        text = uiState.user?.name ?: "User",
-                        style = MaterialTheme.typography.headlineSmall,
-                        fontWeight = FontWeight.Bold,
-                        color = MaterialTheme.colorScheme.primary
-                    )
-
-                    uiState.user?.email?.let { email ->
-                        Spacer(modifier = Modifier.height(2.dp))
-                        Text(
-                            text = email,
-                            style = MaterialTheme.typography.bodyMedium,
-                            color = MaterialTheme.colorScheme.onSurfaceVariant,
-                            maxLines = 1
-                        )
-                    }
-                }
-            }
-        }
-
-        Spacer(modifier = Modifier.height(24.dp))
-
-        // Quick Actions
-        Text(
-            text = "Quick Actions",
-            style = MaterialTheme.typography.headlineSmall,
-            fontWeight = FontWeight.Bold,
-            modifier = Modifier.padding(bottom = 16.dp)
-        )
-
-        // Navigation Buttons
-        onNavigateToTransactions?.let { navigateToTransactions ->
-            ButtonComponent(
-                R.drawable.logo_aichopaicho,
-                "View Transactions",
-                onClick = navigateToTransactions,
-                modifier = Modifier.padding(bottom = 8.dp)
-            )
-        }
-
-        onNavigateToProfile?.let { navigateToProfile ->
-            ButtonComponent(
-                R.drawable.logo_aichopaicho,
-                "Profile Settings",
-                onClick = navigateToProfile,
-                modifier = Modifier.padding(bottom = 8.dp)
-            )
-        }
-
-        ButtonComponent(
-            R.drawable.logo_skip,
-            "Refresh Data",
-            onClick = onRefresh,
-            modifier = Modifier.padding(bottom = 8.dp)
-        )
-
-        Spacer(modifier = Modifier.weight(1f))
-
-        // Sign Out Button
-        onSignOut?.let {
-            ButtonComponent(
-                R.drawable.logo_skip,
-                "Sign Out",
-                onClick = onSignOutClick,
-                modifier = Modifier
-                    .padding(horizontal = 32.dp)
-                    .width(200.dp)
-            )
-        }
-
-        // Error message display
-        uiState.errorMessage?.let { error ->
-            Spacer(modifier = Modifier.height(16.dp))
-            Text(
-                text = error,
-                color = MaterialTheme.colorScheme.error,
-                textAlign = TextAlign.Center,
-                modifier = Modifier.fillMaxWidth()
-            )
-        }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun DashboardScreenPreview() {
-    DashboardScreen(
-        onSignOut = {},
-        onNavigateToProfile = {},
-        onNavigateToTransactions = {}
-    )
 }
